requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==3.0.3
django-core==1.4.1
djangorestframework==3.11.1
factory-boy==3.2.1
Faker==9.8.3
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.2
msgpack==1.0.3
Pillow==8.4.0
prompt-toolkit==3.0.23
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.27.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swagger-client @ file:///Users/vynguyen07/Desktop/Computer_Science/S5/web_dev/week_14/bioweb/python_client
text-unidecode==1.3
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.interface==5.4.0
